+++ src/Cbb/Pitch.cpp
@@ -4,12 +4,12 @@
 namespace Cbb {
 
 
-Pitch::Pitch(const float frequencyHz, const Tuning tuning) noexcept :
+Pitch::Pitch(const double frequencyHz, const Tuning tuning) noexcept :
     Pitch {Midi::frequencyHzToNoteNumber(frequencyHz, tuning)}
 {
 }
 
-float Pitch::getFrequencyHz(const Tuning tuning) const noexcept
+double Pitch::getFrequencyHz(const Tuning tuning) const noexcept
 {
     return Midi::noteNumberToFrequencyHz(getNoteNumber(), tuning);
 }
